+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- `RequestCode::Forward` and `message::send_forward_request`, since forwarding
+  messages between agents is no longer supported.
+
 ## [0.11.0] - 2024-03-25
 
 ### Changed
@@ -263,6 +270,7 @@ without relying on the content of the response.
 
 - `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
+[Unreleased]: https://github.com/petabi/oinq/compare/0.11.0...main
 [0.11.0]: https://github.com/petabi/oinq/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/petabi/oinq/compare/0.9.3...0.10.0
 [0.9.3]: https://github.com/petabi/oinq/compare/0.9.2...0.9.3